requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
anthropic>=0.45.0
chromadb>=0.5.0
langchain-text-splitters>=0.2.0
langchain-community>=0.2.0
pypdf>=4.0.0
docx2txt>=0.8
python-dotenv>=1.0.0
python-multipart>=0.0.9
sentence-transformers>=2.7.0
